backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
Flask==3.1.1
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
flask-cors==6.0.0
fsspec==2025.5.1
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.18.0
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.0rc1
ollama==0.5.1
packaging==23.2
pandas==2.2.2
pillow==11.2.1
portalocker==2.10.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.14.2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==2.7.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
